ORG-121 (User management: enable administrator to assign roles to other users) and ORG-123 (User Management: enable administrator to delete user accounts)

+++ views/jsx/scenarios/ScenarioDeleteButton.jsx
@@ -11,17 +11,20 @@ var Navigation = Router.Navigation;
 var ScenarioDeleteButton = React.createClass({
   mixins: [Navigation, UserHasRoleMixin, UserIsCreatorMixin, FlashQueue.Mixin],
   handleClick: function() {
-    var url = api.reverse('scenario_by_uuid', { uuid : this.props.scenario.uuid });
-    $.ajax(url, {
-      type: 'DELETE',
-      error: this.flashOnAjaxError(url, 'Error trying to delete scenario'),
-      success: (result) => {
-        this.transitionTo('scenarioList');
-        if (typeof this.props.onChange == 'function') {
-          this.props.onChange();
+    var sure = window.confirm('Are you sure you want to delete this version of the scenario?');
+    if (sure) {
+      var url = api.reverse('scenario_by_uuid', { uuid : this.props.scenario.uuid });
+      $.ajax(url, {
+        type: 'DELETE',
+        error: this.flashOnAjaxError(url, 'Error trying to delete scenario'),
+        success: (result) => {
+          this.transitionTo('scenarioList');
+          if (typeof this.props.onChange == 'function') {
+            this.props.onChange();
+          }
         }
-      }
-    });
+      });
+    }
   },
   render: function() {
     if (this.userHasRole('admin') || this.userIsCreator(this.props.scenario)) {
